moved account login and account creation functions to their respective files
